// src/app/core/_models/users/clients/clientesSearch.model.ts
export interface ClienteItem {
    id:                 number;
    nom:            string;
    nit:                number;
    fing:               string;
    ci:                 number;
    ciExpedit:          number;
    nombre:             string;
    app:                string;
    apm:                string;
}